docs(changelog): expand Test Lab decomposition entry with pass-3 details
