+++ docs/changelog.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- **Refactor** — Test Lab page (`frontend/src/pages/TestLab.jsx`) trimmed from 1,907 → ~1,140 lines (audit §3.1). Pass 1 extracted presentational helpers — `ProjIcon`, `PipelinePanel` + `PIPELINE_STAGES`, `LiveLog`, `QueueRow` (under `frontend/src/components/test-lab/`); sessionStorage helpers (`loadPersistedRun` / `persistRun` / `clearPersistedRun` / `STORAGE_KEY` / `LOG_CAP`) into `frontend/src/utils/testLabPersistence.js`; attachment MIME guard + size caps into `frontend/src/utils/testLabAttachments.js`. Pass 2 split the three biggest remaining render blocks: `<TestLabTopBar>` (brand + tablist + Record CTA), `<TestLabProjectSidebar>` (left rail), and `<TestLabRunCenter>` (the ~215-line IIFE for the attached-run middle column — banners, inner-tabs, 3-sub-column pipeline view). `QueueTab` now imports `QueueRow`, `EmptyState`, and the `Clock` icon directly instead of receiving them via DI props. Page render now reads as a state machine — `activeRun ? <RunCenter /> : <ConfigPanel />`. The G9 parallel-runs SSE handler + state Maps remain in-file pending that migration's completion. No behaviour change — pure decomposition.
+- **Refactor** — Test Lab page (`frontend/src/pages/TestLab.jsx`) trimmed from 1,907 → ~995 lines (~48% reduction, audit §3.1). Pass 1 extracted presentational helpers — `ProjIcon`, `PipelinePanel` + `PIPELINE_STAGES`, `LiveLog`, `QueueRow` (under `frontend/src/components/test-lab/`); sessionStorage helpers (`loadPersistedRun` / `persistRun` / `clearPersistedRun` / `STORAGE_KEY` / `LOG_CAP`) into `frontend/src/utils/testLabPersistence.js`; attachment MIME guard + size caps into `frontend/src/utils/testLabAttachments.js`. Pass 2 split the three biggest remaining render blocks: `<TestLabTopBar>`, `<TestLabProjectSidebar>`, and `<TestLabRunCenter>` (the ~215-line IIFE for the attached-run middle column). Pass 3 extracted three logic seams into their own modules: `useTestLabAttachments` (file upload + Jira import state machine + binary-detection / MIME guard), `useProjectEnvironments` (DIF-012 per-project env list fetch with viewer-403 swallow + project-switch reset), and `splitGeneratedOutcome` (the drafts-vs-auto-approved derivation that was a 30-line `useMemo` body). `ProjIcon` avatar palette now driven by a `--avatar-h` CSS custom property — no raw HSL strings inline (STANDARDS.md:383). `QueueTab` imports `QueueRow`, `EmptyState`, and `Clock` directly instead of receiving them via DI props. Page render now reads as a state machine — `activeRun ? <RunCenter /> : <ConfigPanel />`. The G9 parallel-runs SSE handler + state Maps remain in-file pending that migration's completion. No behaviour change — pure decomposition.
 
 - **SEC / Perf** — Web fonts (Inter, JetBrains Mono) self-hosted via `@fontsource` instead of `@import url('https://fonts.googleapis.com/...')`. The Google Fonts CDN saw every visitor's IP + User-Agent on every page load, blocking GDPR-compliant EU deployments (CJEU C-101/01 / Munich VG 2022) and preventing CSP hardening of `style-src` / `font-src` beyond `'self' data:`. Also removes a ~120ms render-blocking external request on cold loads and makes the app work in air-gapped enterprise deployments. Imports added to `frontend/src/main.jsx`; bundle size impact is offset by eliminating the external stylesheet round-trip.
